requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.21
pydantic>=2.0.0
panel>=1.3.8
param>=1.13.0
python-dotenv>=1.0.0
uvicorn>=0.24.
fastapi
langchain-google-genai
langchain
chromadb
pypdf
